move more logic up and outta doctor-general-cli
