+++ packages/doctor-general/constants.js
@@ -1,1 +1,2 @@
 export const MAGIC_IMPORT_KEY = 'drgen-import-above'
+export const TESTABLE_EXAMPLE = 'test=true'
